Fix WindRose function
