+++ react/components/station_dashboard/wind_rose_utils.js
@@ -1,3 +1,36 @@
+import React, { useEffect, useRef, useState, useMemo } from 'react';
+import { get_station_field_data, get_station_field_units } from '../utils/station_data_format_util';
+
+export const windSpeedBins = [
+  { min: 0, max: 8, label: "=< 8 m/s" },
+  { min: 8, max: 16, label: "9-16 m/s" },
+  { min: 16, max: 24, label: "17-24 m/s" },
+  { min: 24, max: 32, label: "25-32 m/s" },
+  { min: 32, max: 40, label: "33-40 m/s" },
+  { min: 40, max: 48, label: "41-48 m/s" },
+  { min: 48, max: 56, label: "49-56 m/s" },
+  { min: 56, max: 64, label: "57-64 m/s" },
+  { min: 64, max: undefined, label: ">= 65 m/s" },
+
+];
+export const colorPalette= [
+  '#5E4FA2', // Deep Purple
+'#3288BD', // Blue
+'#66C2A5', // Teal
+'#ABDDA4', // Light Green
+'#E6F598', // Yellow-Green
+'#FEE08B', // Yellow
+'#FDAE61', // Orange
+'#F46D43', // Red-Orange
+'#D53E4F', // Deep Red
+];
+
+export const cardinalPoints = [
+  'N', 'NNE', 'NE', 'ENE', 'E', 'ESE', 'SE', 'SSE', 'S', 'SSW', 'SW', 'WSW', 'W', 'WNW', 'NW', 'NNW',
+];
+
+
+
 export function categorizeWindDirection(direction) {
   const cardinalPoints = [
     'N', 'NNE', 'NE', 'ENE', 'E', 'ESE', 'SE', 'SSE', 'S', 'SSW', 'SW', 'WSW', 'W', 'WNW', 'NW', 'NNW',
@@ -112,4 +145,51 @@ export function parseWindData(freqFrac, windSpdLabel){
     //console.log(data)
     return data;
 
+}
+
+function ChartComponent({ data, chartKey }) {
+  const chartContainerRef = useRef(null);
+
+  useEffect(() => {
+    const chart = new Chart({
+      container: chartContainerRef.current});
+    chart.options({
+      type: "interval",
+      title:{
+        title: chartKey,
+      },
+      autoFit: true,
+      height: 300,
+      padding: 40,
+      data: data,
+      encode: { x: "direction", y: "value", color: "windSpeedBin", size: 18 },
+      transform: [{ type: "stackY" }],
+      scale: {
+          color: {
+          range: colorPalette,
+          },
+      },
+      coordinate: { type: "polar" },
+      axis: {
+          x: { line: true, grid: true, gridLineDash: [0, 0], gridLineWidth: 1 },
+          y: { title: false, line: true, gridLineWidth: 1 },
+      },
+      tooltip: {
+          title: (d) => d.direction,
+          items: [
+          (d, i, data, column) => ({
+              name: d.windSpeedBin,
+              value: d.value,
+              channel: "y",
+          }),
+          ],
+      },
+      interaction: { tooltip: { shared: true } },
+      });
+    chart.render();
+
+    return () => chart.destroy(); // Cleanup on unmount
+  }, [data]);
+
+  return <div id={`chart-${chartKey}`} ref={chartContainerRef} className="chart-container" />;
 }
